Move context into a separate folder, add PageContext
