+++ src/User.js
@@ -5,7 +5,7 @@ import classNames from "classnames";
 import PropTypes from "prop-types";
 import React from "react";
 
-import AdminContext from "./context";
+import AdminContext from "./context/AdminContext";
 import MenuItem from "./MenuItem";
 
 const styles = theme => ({
