+++ src/context/PageContext.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const PageContext = React.createContext({
+  title: undefined,
+  route: undefined,
+  logger: undefined,
+  data: undefined,
+});
+
+export const usePage = () => React.useContext(PageContext);
+export default PageContext;
