+++ src/pages/ErrorPage.js
@@ -3,7 +3,7 @@ import { withStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import React from "react";
 
-import AdminContext from "../context";
+import AdminContext from "../context/AdminContext";
 import { t, Translate } from "../utils";
 import CardPage from "./CardPage";
