+++ src/pages/MePage.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import React from "react";
 
 import Content from "../Content";
-import AdminContext from "../context";
+import AdminContext from "../context/AdminContext";
 import TitleBar from "../TitleBar";
 import ChangePasswordForm from "./ChangePasswordForm";
 import SettingsForm from "./SettingsForm";
